move context.hawk into separate file
